chore: move backend files to /scripts/
