+++ scripts/getCurrentTime.js
