+++ scripts/readFile.js
